@dougwilson 2015-05-20
So I know a lot has changed in this module since the PR, but I have picked the PR into a branch in the repo (https://github.com/jshttp/cookie/tree/mscdex/pref-changes) that I have been rebasing. I've also picked out at least one change directly for a patch release.